Cast size_t when passing to is_string in pam-util/vector test suite

+++ tests/pam-util/vector-t.c
@@ -62,9 +62,10 @@ main(void)
     is_int(4, copy->count, "...and has right count");
     is_int(4, copy->allocated, "...and has right allocated count");
     for (i = 0; i < 4; i++) {
-        is_string(cstring, copy->strings[i], "...and string %d is right", i);
+        is_string(cstring, copy->strings[i], "...and string %lu is right",
+                  (unsigned long) i);
         ok(copy->strings[i] != vector->strings[i],
-           "...and pointer %d is different", i);
+           "...and pointer %lu is different", (unsigned long) i);
     }
     vector_free(copy);
     vector_clear(vector);
